execution-zones/docs-agent/src/requirements.txt
strands-agents[a2a,otel]~=1.25.0
aws-opentelemetry-distro~=0.10.0
uvicorn~=0.34.0
fastapi~=0.115.0
boto3~=1.42.0
requests~=2.31.0
